pyproject.toml
[build-system]
requires = [
    "setuptools>=64",
    "setuptools_scm>=8",
    "pyqt-distutils>=0.7.3",
    "wheel>=0.37.0",
    "dulwich",
    "randovania-scm-version-schema",
]
build-backend = "setuptools.build_meta"

[project]
name = "randovania"
authors = [
    {name = "Henrique Gemignani Passos Lima"},
]
description = "A randomizer platform for a multitude of games."
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Environment :: Win32 (MS Windows)",
    "Environment :: X11 Applications :: Qt",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Topic :: Games/Entertainment",
]
requires-python = ">=3.11"
dynamic = ["version"]

dependencies = [
    # server communication, binary version of logic db
    "construct",

    # Default name for exporting presets
    "python-slugify",

    # permalinks
    "bitstruct",

    # obfuscator, guest login, server code
    "cryptography",

    # Generator
    "networkx",
    "tenacity>=7.0.0",

    # server communication
    "python-socketio[asyncio_client]",
    "aiohttp>=3.9.3",

    # nonblocking file IO
    "aiofiles",

    # Immutable dicts, used in many places
    "frozendict",

    # Versioning presets
    "dulwich>=0.20",

    # error reports
    "sentry-sdk",

    # Patch GamesPatches JSON from rdvgame
    "json-delta"
]

[project.optional-dependencies]
exporters = [
    # AM2R
    "am2r-yams>=2.4.0",

    # Cave Story
    "cave-story-randomizer>=2.3.3",
    "tsc-utils>=0.2.5",

    # Factorio
    "factorio-randovania-mod>=0.4.0",

    # Metroid Dread
    "open-dread-rando>=2.11.0",

    # Metroid Prime 1
    "py_randomprime>=1.24.5",
    "random-enemy-attributes>=1.0.3",

    # Metroid Prime 1/2
    "ppc-asm>=1.2.1",
    "open-prime-rando>=0.11.2",
    "retro-data-structures>=0.24.0",

    # Metroid Prime 2
    "mp2hudcolor>=1.0.10",

    # Metroid: Samus Returns
    "open-samus-returns-rando>=1.3.3",

    # Super Metroid
    "SuperDuperMetroid>=1.2.0",
]

gui = [
    "PySide6-Essentials>=6.5",
    "pyqtdarktheme-fork>=2.3.0",
    "markdown",
    "qasync",

    "appdirs",
    # Minimum version compatible with PySide 6.4. 3.6 is also needed for Python 3.11
    "matplotlib>=3.6.2",

    "nod>=1.9",
    "async-wiiload",
    "dolphin-memory-engine>=1.0.2",
    "pid>=3.0.0",
]

server = [
    "sentry-sdk[flask]",
    "flask-discord",
    "flask-socketio",
    "prometheus-flask-exporter",
    "eventlet",
    "peewee",
    "requests-oauthlib",
    "cachetools",  # for TTLCache

    "py-cord>=2.3.1",  # for randovania/server/discord; 2.3.1+ is needed for Python 3.11
    "Pillow>=9.0.0",  # for randovania/server/discord/database_command
    "graphviz",  # for randovania/server/discord/database_command
]

test = [
    "pytest",
    "pytest-cov",
    "pytest-qt",
    "pytest-asyncio",
    "pytest-mock",
    "pytest-localftpserver",
    "pytest-xdist[psutil]",
]

typing = [
    "mypy",
    "construct-typing",
    "types-aiofiles",
    "types-peewee",
]

website = [
    "htmlmin",
    "ruamel.yaml",
]

[project.readme]
file = "README.md"
content-type = "text/markdown"

[project.urls]
Homepage = "https://github.com/randovania/randovania"

[project.entry-points.console_scripts]
randovania = "randovania.__main__:main"

[tool.setuptools.packages.find]
include = ["randovania*"]

[tool.setuptools_scm]
local_scheme = "no-local-version"
version_file = "randovania/version.py"
version_scheme = "randovania_scm_version_schema:version_scheme"
git_describe_command = [
    "git",
    "describe",
    "--dirty",
    "--tags",
    "--long",
    "--match",
    "v[0-9]*",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = [
    "E", "F", "W", "C90", "I", "UP", "C4",
    "RSE",
    "TCH",
    "PTH",
    "COM818", "COM819",
    "ISC",
    "PIE",
    "PT",

    "PLC",
    "PLC0208",  # iterating over set is also not deterministic, so we shouldn't do that!

    "PLE",
    "PLR0402", "PLR1701", "PLR1711", "PLR1722", "PLR0206",
    "PLR0133", "PLR0124",
    "PLW",
]
extend-ignore = [
    "ISC001",  # may cause conflicts with formatter
    "PLW0603",  # we use some globals here and there
    "PLW2901",  # TODO: should actually fix these
]

[tool.ruff.lint.isort]
# This is very desirable, but causes issues with py-cord and some usages of construct_pack.encode
# required-imports = ["from __future__ import annotations"]

[tool.ruff.lint.mccabe]
# Flag errors (`C901`) whenever the complexity level exceeds 25.
# Defaults to 10, but we're being very flexible right now
max-complexity = 25

[tool.ruff.lint.flake8-type-checking]
runtime-evaluated-base-classes = [
    "randovania.bitpacking.bitpacking.BitPackDataclass",
    "randovania.bitpacking.type_enforcement.DataclassPostInitTypeCheck",
    "randovania.bitpacking.json_dataclass.JsonDataclass",
    "randovania.layout.base.base_configuration.BaseConfiguration",
    "randovania.exporter.game_exporter.GameExportParams",
]

[tool.mypy]
files = [
    "randovania/bitpacking/",
    "randovania/cli/__init__.py",
    "randovania/game_connection/connector/dread_remote_connector.py",
    "randovania/game_connection/connector/mercury_remote_connector.py",
    "randovania/game_connection/connector/msr_remote_connector.py",
    "randovania/game_connection/executor/dread_executor.py",
    "randovania/game_connection/executor/msr_executor.py",
    "randovania/cli/commands/new_game.py",
    "randovania/game_description/",
    "randovania/games/am2r/gui/dialog/cosmetic_patches_dialog.py",
    "randovania/games/blank/",
    "randovania/games/cave_story/gui/dialog/cs_cosmetic_patches_dialog.py",
    "randovania/games/dread/gui/dialog/dread_cosmetic_patches_dialog.py",
    "randovania/games/prime1/gui/dialog/prime_cosmetic_patches_dialog.py",
    "randovania/games/prime2/exporter/patch_data_factory.py",
    "randovania/games/prime2/gui/dialog/echoes_cosmetic_patches_dialog.py",
    "randovania/games/prime2/layout/echoes_user_preferences.py",
    "randovania/games/prime3/",
    "randovania/games/samus_returns/",
    "randovania/games/super_metroid/gui/dialog/super_cosmetic_patches_dialog.py",
    "randovania/generator/",
    "randovania/gui/docks/resource_database_editor.py",
    "randovania/gui/dialog/base_cosmetic_patches_dialog.py",
    "randovania/gui/dialog/connections_editor.py",
    "randovania/gui/preset_settings/trick_level_tab.py",
    "randovania/gui/lib/signal_handling.py",
    "randovania/gui/widgets/node_selector_widget.py",
    "randovania/lib/",
    "test/bitpacking/",
    "test/game_description/",
    "test/games/blank/",
    "test/games/cave_story/gui/test_cs_cosmetic_patches_dialog.py",
    "test/games/dread/gui/test_dread_cosmetic_patches_dialog.py",
    "test/games/prime1/gui/dialog/test_prime_cosmetic_patches_dialog.py",
    "test/games/prime2/gui/dialog/test_echoes_cosmetic_patches_dialog.py",
    "test/games/prime3/gui/dialog/test_corruption_cosmetic_patches_dialog.py",
    "test/generator/",
    "test/lib/",
    "test/gui/dialog/test_connections_editor.py",
    "test/gui/widgets/test_node_selector_widget.py",
]
follow_imports = "silent"
disallow_untyped_defs = true
local_partial_types = true

[[tool.mypy.overrides]]
module = "test.*"
disallow_untyped_defs = false
check_untyped_defs = true
disable_error_code = "method-assign"

[[tool.mypy.overrides]]
module = "randovania.games.*.gui.generated.*"
disallow_untyped_defs = false
